Fix misleading examples in README
According to the original example, calling the function
`lights.set_pixel(...)` should change the light of the LED in the given
position. However, this function only sets the internal buffer, and to
actually change the values of the LEDs a call to `lights.show()` should
be done.

The commit updates the documentation in the README files (both README.md
and library/README.md) with more precise description.

The commit fixes #6.